WIP refactor and add AddUsers to twitterclient
