+++ cmd/apply.go
@@ -8,12 +8,12 @@ import (
 	"strings"
 
 	"github.com/amitizle/twitter_lists_manager/internal/printer"
-	"github.com/amitizle/twitter_lists_manager/internal/twitter_client"
+	"github.com/amitizle/twitter_lists_manager/internal/twitterclient"
 )
 
 // ApplyLists applying the JSON formatted inFile, modifying, adding and removing lists and
 // members from the lists
-func ApplyLists(client *twitter_client.Client, inFile string) {
+func ApplyLists(client *twitterclient.Client, inFile string) {
 	localLists, err := readJSONLists(inFile)
 	if err != nil {
 		printer.Fatalf("Error while reading lists from JSON file %s: %v", inFile, err)
@@ -24,8 +24,8 @@ func ApplyLists(client *twitter_client.Client, inFile string) {
 		printer.Fatalf("Error: %v", err)
 	}
 	printer.Infof("%#v", ownedLists)
-	localListsMap := make(map[string]*twitter_client.List)
-	ownedListsMap := make(map[string]*twitter_client.List)
+	localListsMap := make(map[string]*twitterclient.List)
+	ownedListsMap := make(map[string]*twitterclient.List)
 	for _, list := range localLists {
 		localListsMap[list.Slug] = &list // TODO something?
 	}
@@ -39,7 +39,7 @@ func ApplyLists(client *twitter_client.Client, inFile string) {
 			screenNamesToAdd := sliceDifference(localList.Members, ownedList.Members)
 			printer.Infof(strings.Join(screenNamesToAdd, "\n"))
 			printer.Redf(strings.Join(screenNamesToRemove, "\n"))
-			// diff members
+			client.AddUsersToList(ownedList, screenNamesToAdd)
 			// update list https://developer.twitter.com/en/docs/accounts-and-users/create-manage-lists/api-reference/post-lists-update
 		} else { // new list TODO support deleting lists
 		}
@@ -63,27 +63,27 @@ func sliceDifference(a, b []string) []string {
 	return diff
 }
 
-func readJSONLists(jsonFilePath string) ([]twitter_client.List, error) {
+func readJSONLists(jsonFilePath string) ([]twitterclient.List, error) {
 	fullPath, err := filepath.Abs(jsonFilePath)
 	if err != nil {
-		return []twitter_client.List{}, err
+		return []twitterclient.List{}, err
 	}
 	jsonFile, err := os.Open(fullPath)
 	if err != nil {
-		return []twitter_client.List{}, err
+		return []twitterclient.List{}, err
 	}
 
 	defer jsonFile.Close()
 
 	byteValue, err := ioutil.ReadAll(jsonFile)
 	if err != nil {
-		return []twitter_client.List{}, err
+		return []twitterclient.List{}, err
 	}
 
-	lists := make([]twitter_client.List, 0)
+	lists := make([]twitterclient.List, 0)
 	err = json.Unmarshal(byteValue, &lists)
 	if err != nil {
-		return []twitter_client.List{}, err
+		return []twitterclient.List{}, err
 	}
 
 	return lists, nil
