+++ cmd/import.go
@@ -6,11 +6,11 @@ import (
 	"os"
 
 	"github.com/amitizle/twitter_lists_manager/internal/printer"
-	"github.com/amitizle/twitter_lists_manager/internal/twitter_client"
+	"github.com/amitizle/twitter_lists_manager/internal/twitterclient"
 )
 
 // ImportLists imports the lists owned by the authenticated user to a JSON formatted outputFile
-func ImportLists(client *twitter_client.Client, outputFile string) { // TODO rename to export
+func ImportLists(client *twitterclient.Client, outputFile string) { // TODO rename to export
 	ownedLists, err := client.GetOwnedLists(nil)
 	if err != nil {
 		printer.Fatalf("Error: %v", err)
