+++ example.json
@@ -1,19 +1,20 @@
 [
-  {
-    "Slug": "programming",
-    "Name": "Programming",
-    "Description": "",
-    "Members": [
-      "golang",
-      "joeerl",
-      "googledevs",
-      "peterseibel",
-      "spolsky",
-      "rvirding",
-      "jwz",
-      "shanselman",
-      "bradfitz"
-    ],
-    "Mode": "private"
-  }
+    {
+        "Slug": "programming",
+        "Name": "Programming",
+        "Description": "",
+        "Members": [
+            "golang",
+            "joeerl",
+            "googledevs",
+            "peterseibel",
+            "spolsky",
+            "rvirding",
+            "jwz",
+            "shanselman",
+            "bradfitz",
+            "HashiCorp"
+        ],
+        "Mode": "private"
+    }
 ]
